rename stats components file
