+++ lib/src/presentation/components/stats.dart
