+++ lib/src/presentation/overview.dart
@@ -6,7 +6,7 @@ import 'package:geo_steps/src/presentation/components/calender.dart';
 import 'package:geo_steps/src/presentation/components/icons.dart';
 import 'package:geo_steps/src/presentation/components/lines.dart';
 import 'package:geo_steps/src/presentation/components/map.dart';
-import 'package:geo_steps/src/presentation/components/overview_stats.dart';
+import 'package:geo_steps/src/presentation/components/stats.dart';
 import 'package:geo_steps/src/utils/datetime.dart';
 import 'package:geo_steps/src/utils/sizing.dart';
 import 'package:geolocator/geolocator.dart';
